Schedule functionality
When a user creates an event it gets put on the right day in the calendar. the time is buggy and you need to reenter if you edit and it doesnt show properly in the dalender but you can click on events and even drag it in the time of day to change it and drag out the bottom to make it longer. I changed the time functionality to a seperate date/starting time/ ending time . You will need to make migrations, it will be easier if you delete all events before but it might as you for a auto enter for date when you migrate but just put "timezone.now"